Fix grep Command, a Couple of Comments, and Readability
This commit fixes the grep command to use the correct option "l" instead of "1". Realizing how similar the two characters are next to each other, I ended up also distinguishing all of the options with separate minus signs. In addition, I modified some comments to reflect the current state of the code. grep is now working properly. It also no longer has option f, which was added and removed during error diagnosis. The part of a comment related to it has been removed.